+++ src/Compilers/Core/Portable/Collections/DictionaryExtensions.cs
@@ -2,7 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Collections.Generic;
+using Microsoft.CodeAnalysis.PooledObjects;
 
 namespace Microsoft.CodeAnalysis
 {
@@ -48,5 +50,29 @@ namespace Microsoft.CodeAnalysis
             return true;
         }
 #endif
+
+        public static void AddToAccumulator<TKey, T>(this Dictionary<TKey, object> accumulator, T item, Func<T, TKey> keySelector)
+            where TKey : notnull
+            where T : notnull
+        {
+            var key = keySelector(item);
+            if (accumulator.TryGetValue(key, out var existingValueOrArray))
+            {
+                if (existingValueOrArray is not ArrayBuilder<T> arrayBuilder)
+                {
+                    // Just a single value in the accumulator so far.  Convert to using a builder.
+                    arrayBuilder = ArrayBuilder<T>.GetInstance(capacity: 2);
+                    arrayBuilder.Add((T)existingValueOrArray);
+                    accumulator[key] = arrayBuilder;
+                }
+
+                arrayBuilder.Add(item);
+            }
+            else
+            {
+                // Nothing in the dictionary so far.  Add the item directly.
+                accumulator.Add(key, item);
+            }
+        }
     }
 }
